Most postman passing

+++ keepr/Services/VaultKeepsService.cs
@@ -12,18 +12,32 @@ namespace keepr.Services
             _repo = repo;
         }
 
-        internal VaultKeep Create(VaultKeep vaultKeepData)
+        internal VaultKeep Create(VaultKeep vaultKeepData, string userId)
         {
-            return _repo.Create(vaultKeepData);
+            VaultKeep newVaultKeep = _repo.Create(vaultKeepData);
+
+            if (newVaultKeep.CreatorId != userId)
+            {
+                throw new System.Exception("You do not have the right to post this.");
+            }
+            else
+            {
+
+                return newVaultKeep;
+            }
         }
         internal VaultKeep Get(int id)
         {
             return _repo.Get(id);
         }
 
-        internal void Delete(int id)
+        internal void Delete(int id, string userId)
         {
-            Get(id);
+            VaultKeep found = Get(id);
+            if (found.CreatorId != userId)
+            {
+                throw new System.Exception("You do not have the right to delete this.");
+            }
             _repo.Delete(id);
         }
     }
